Rename cmd/archer package to cmd/ecs-preview
