+++ cmd/ecs-preview/template/template_windows.go
