+++ internal/pkg/cli/pipeline.go
@@ -4,7 +4,7 @@
 package cli
 
 import (
-	"github.com/aws/amazon-ecs-cli-v2/cmd/archer/template"
+	"github.com/aws/amazon-ecs-cli-v2/cmd/ecs-preview/template"
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/cli/group"
 	"github.com/spf13/cobra"
 )
