+++ internal/pkg/cli/version.go
@@ -7,7 +7,7 @@ package cli
 import (
 	"fmt"
 
-	"github.com/aws/amazon-ecs-cli-v2/cmd/archer/template"
+	"github.com/aws/amazon-ecs-cli-v2/cmd/ecs-preview/template"
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/cli/group"
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/version"
 	"github.com/spf13/cobra"
